clean selection of device and dtype
