+++ crates/kornia-vlm/src/paligemma/mod.rs
@@ -81,18 +81,6 @@ pub struct Paligemma {
 }
 
 impl Paligemma {
-    #[cfg(feature = "cuda")]
-    fn cuda_supports_bf16(device_id: usize) -> bool {
-        use cudarc::driver::{sys::CUdevice_attribute, CudaDevice};
-        if let Ok(dev) = CudaDevice::new(device_id) {
-            if let Ok(major) =
-                dev.attribute(CUdevice_attribute::CU_DEVICE_ATTRIBUTE_COMPUTE_CAPABILITY_MAJOR)
-            {
-                return major >= 8;
-            }
-        }
-        false     
-    }
     /// Create a new Paligemma model
     ///
     /// # Arguments
@@ -101,30 +89,8 @@ impl Paligemma {
     ///
     /// # Returns
     pub fn new(config: PaligemmaConfig) -> Result<Self, PaligemmaError> {
-        #[cfg(feature = "cuda")]
-        let (device, dtype) = match Device::cuda_if_available(0) {
-            Ok(device) => {
-                let dtype = if Self::cuda_supports_bf16(0) {
-                    log::info!("GPU supports BF16, using BF16");
-                    DType::BF16
-                } else {
-                    log::warn!(
-                        "GPU does not support BF16, falling back to FP16. \
-                        Note: FP16 may overflow in attention logits and produce lower-quality results."
-                    );
-                    DType::F16
-                };
-                (device, dtype)
-            }
-            Err(e) => {
-                log::warn!("CUDA not available, defaulting to CPU: {e:?}");
-                (Device::Cpu, DType::F32)
-            }
-        };
-
-        #[cfg(not(feature = "cuda"))]
-        let (device, dtype) = (Device::Cpu, DType::F32);
-
+        use crate::device::get_device_and_dtype;
+        let (device, dtype) = get_device_and_dtype();
         let (model, tokenizer) = Self::load_model(dtype, &device)?;
         let img_buf = Image::from_size_val([224, 224].into(), 0, CpuAllocator)?;
         let pipeline = TextGeneration::new(model, tokenizer, device, config.into());
