+++ crates/kornia-vlm/src/smolvlm2/mod.rs
@@ -137,18 +137,6 @@ pub struct SmolVlm2<const N: usize, A: ImageAllocator> {
 }
 
 impl<const N: usize, A: ImageAllocator> SmolVlm2<N, A> {
-    #[cfg(feature = "cuda")]
-    fn cuda_supports_bf16(device_id: usize) -> bool {
-        use cudarc::driver::{sys::CUdevice_attribute, CudaDevice};
-        if let Ok(dev) = CudaDevice::new(device_id) {
-            if let Ok(major) =
-                dev.attribute(CUdevice_attribute::CU_DEVICE_ATTRIBUTE_COMPUTE_CAPABILITY_MAJOR)
-            {
-                return major >= 8;
-            }
-        }
-        false
-    }
     const MODEL_IDENTIFIER: &'static str = "HuggingFaceTB/SmolVLM2-2.2B-Instruct";
     const IMG_PROCESSOR_CONFIG: ImageProcessorConfig = ImageProcessorConfig {
         size_longest_edge: 1536,
@@ -173,29 +161,8 @@ impl<const N: usize, A: ImageAllocator> SmolVlm2<N, A> {
     /// Create a new SmolVLM2 instance with the default configuration.
     /// This will download weights from HuggingFace Hub.
     pub fn new(config: SmolVlm2Config) -> Result<Self, SmolVlm2Error> {
-        #[cfg(feature = "cuda")]
-        let (device, dtype) = match Device::cuda_if_available(0) {
-            Ok(device) => {
-                let dtype = if Self::cuda_supports_bf16(0) {
-                    log::info!("GPU supports BF16, using BF16");
-                    DType::BF16
-                } else {
-                    log::warn!(
-                        "GPU does not support BF16, falling back to FP16. \
-                        Note: FP16 may overflow in attention logits and produce lower-quality results."
-                    );
-                    DType::F16
-                };
-                (device, dtype)
-            }
-            Err(e) => {
-                log::warn!("CUDA not available, defaulting to CPU: {e:?}");
-                (Device::Cpu, DType::F32)
-            }
-        };
-
-        #[cfg(not(feature = "cuda"))]
-        let (device, dtype) = (Device::Cpu, DType::F32);
+        use crate::device::get_device_and_dtype;
+        let (device, dtype) = get_device_and_dtype();
 
         let (model, txt_processor, img_processor, vid_processor) =
             Self::load_model(&config, dtype, &device)?;
